+++ PlaneAlerter/Forms/Condition Editor Dialog.resx
