+++ PlaneAlerter/Forms/Condition Editor.resx
