+++ PlaneAlerter/Forms/PinPromptDialog.cs
@@ -1,14 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace PlaneAlerter {
+namespace PlaneAlerter.Forms {
 	public partial class PinPromptDialog : Form {
 		public string PIN { get; set; }
