+++ PlaneAlerter/Forms/PinPromptDialog.resx
