+++ PlaneAlerter/Forms/PropertyInfoForm.resx
