+++ PlaneAlerter/Forms/SettingsForm.cs
@@ -1,11 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Windows.Forms;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Windows.Forms;
 
-namespace PlaneAlerter {
+namespace PlaneAlerter.Forms {
 	/// <summary>
 	/// Form for changing settings
 	/// </summary>
