+++ PlaneAlerter/Forms/SettingsForm.resx
