+++ PlaneAlerter/Program.cs
@@ -11,7 +11,7 @@ namespace PlaneAlerter {
 			Application.EnableVisualStyles();
 			Application.SetHighDpiMode(HighDpiMode.SystemAware);
 			Application.SetCompatibleTextRenderingDefault(false);
-			Application.Run(new PlaneAlerter());
+			Application.Run(new Forms.PlaneAlerter());
 		}
 	}
 }
